+++ contentful_management/client.py
@@ -16,10 +16,12 @@ from .uploads_proxy import UploadsProxy
 from .api_keys_proxy import ApiKeysProxy
 from .webhooks_proxy import WebhooksProxy
 from .snapshots_proxy import SnapshotsProxy
+from .api_usages_proxy import ApiUsagesProxy
 from .environments_proxy import EnvironmentsProxy
 from .webhooks_call_proxy import WebhooksCallProxy
 from .ui_extensions_proxy import UIExtensionsProxy
 from .content_types_proxy import ContentTypesProxy
+from .usage_periods_proxy import UsagePeriodsProxy
 from .organizations_proxy import OrganizationsProxy
 from .webhooks_health_proxy import WebhooksHealthProxy
 from .preview_api_keys_proxy import PreviewApiKeysProxy
@@ -183,6 +185,40 @@ class Client(object):
         """
 
         return OrganizationsProxy(self)
+
+    def usage_periods(self, organization_id):
+        """
+        Provides access to an organizations usage periods.
+
+        API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/usage-periods
+
+        :return: :class:`UsagePeriodsProxy <contentful_management.usage_periods_proxy.UsagePeriodsProxy>` object.
+        :rtype: contentful.usage_periods_proxy.UsagePeriodsProxy
+
+        Usage:
+
+            >>> usage_periods_proxy = client.usage_periods('organization_id')
+            <UsagePeriodsProxy organization_id='organization_id'>
+        """
+
+        return UsagePeriodsProxy(self, organization_id)
+
+    def api_usage(self, organization_id):
+        """
+        Provides access to an organizations api usage metrics.
+
+        API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/api-usages
+
+        :return: :class:`ApiUsagesProxy <contentful_management.api_usages_proxy.ApiUsagesProxy>` object.
+        :rtype: contentful.api_usages_proxy.ApiUsagesProxy
+
+        Usage:
+
+            >>> api_usages_proxy = client.api_usage('organization_id')
+            <ApiUsagesProxy organization_id='organization_id'>
+        """
+
+        return ApiUsagesProxy(self, organization_id)
 
     def users(self):
         """