+++ contentful_management/resource_builder.py
@@ -9,10 +9,12 @@ from .upload import Upload
 from .api_key import ApiKey
 from .webhook import Webhook
 from .snapshot import Snapshot
+from .api_usage import ApiUsage
 from .environment import Environment
 from .ui_extension import UIExtension
 from .content_type import ContentType
 from .webhook_call import WebhookCall
+from .usage_period import UsagePeriod
 from .organization import Organization
 from .webhook_health import WebhookHealth
 from .preview_api_key import PreviewApiKey
@@ -62,6 +64,8 @@ class ResourceBuilder(object):
             'Space': Space,
             'SpaceMembership': SpaceMembership,
             'Organization': Organization,
+            'UsagePeriod': UsagePeriod,
+            'ApiUsage': ApiUsage,
             'ApiKey': ApiKey,
             'PreviewApiKey': PreviewApiKey,
             'PersonalAccessToken': PersonalAccessToken,