+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Unreleased
+### Added
+* Added Usage Periods API
+* Added API Usages API
 
 ## v2.5.0